refactor: renaming testConnection and healthController, and applying named export
